requirements.txt
pandas>=2.0
numpy>=1.26
matplotlib>=3.8
streamlit>=1.28
anthropic>=0.18
fpdf2>=2.7
reportlab>=4.0
openai>=1.0
pdf2image>=1.17
pillow>=10.0
pymupdf>=1.24
opencv-python-headless>=4.10
pytesseract>=0.3
shapely>=2.0
networkx>=3.2
scipy>=1.10
requests>=2.31
xgboost>=2.0
scikit-learn>=1.4
joblib>=1.3
